requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
docopt==0.6.2
num2words==0.5.10
Pillow==8.0.1
pytz==2020.5
sqlparse==0.4.1
timeago==1.0.15
whitenoise==5.2.0
